# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST Api for covid-19 cases",
        "title": "GO-COVID19 API",
        "contact": {
            "name": "Jezer Dave Bacquian",
            "url": "https://github.com/jezerdave",
            "email": "jezerdavebacquian@gmail.com"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "go-covid19.sideprojects.fun",
    "basePath": "/api/v1",
    "paths": {
        "/countries": {
            "get": {
                "description": "get current covid data",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all countries data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/worldometer.CountryStats"
                            }
                        }
                    }
                }
            }
        },
        "/countries/{country}": {
            "get": {
                "description": "find covid19 related data by country",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "country name / code (philippines/ph/608/phl)",
                        "name": "country",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/worldometer.CountryStats"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.BasicError"
                        }
                    }
                }
            }
        },
        "/doh/ph": {
            "get": {
                "description": "get doh philippines official data from https://ncovtracker.doh.gov.ph/",
                "produces": [
                    "application/json"
                ],
                "summary": "GET DOH PHILIPPINES DATA",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/philippines.StatsAttributes"
                            }
                        }
                    }
                }
            }
        },
        "/doh/ph/hospital-pui": {
            "get": {
                "description": "get doh philippines official data (HOSPITAL PUIs) from https://ncovtracker.doh.gov.ph/",
                "produces": [
                    "application/json"
                ],
                "summary": "GET DOH PHILIPPINES DATA (HOSPITAL PUIs)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/philippines.HsPUIsAttributes"
                            }
                        }
                    }
                }
            }
        },
        "/histories": {
            "get": {
                "description": "get histories",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all countries historical data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/who.HistoryData"
                            }
                        }
                    }
                }
            }
        },
        "/histories/{country}": {
            "get": {
                "description": "find covid19 related historical data by country",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Country Histories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "country name / code (philippines/ph/608/phl)",
                        "name": "country",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/who.HistoryData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.BasicError"
                        }
                    }
                }
            }
        },
        "/states": {
            "get": {
                "description": "get current US-States covid data",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all US-States data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/worldometer.StatesStats"
                            }
                        }
                    }
                }
            }
        },
        "/states/{state}": {
            "get": {
                "description": "find covid19 related data by state",
                "produces": [
                    "application/json"
                ],
                "summary": "Find State",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State Name / State Abbreviation",
                        "name": "state",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/worldometer.StatesStats"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.BasicError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "jsons.Country": {
            "type": "object",
            "properties": {
                "cca2": {
                    "type": "string"
                },
                "cca3": {
                    "type": "string"
                },
                "ccn3": {
                    "type": "string"
                },
                "cioc": {
                    "type": "string"
                },
                "name": {
                    "type": "object",
                    "$ref": "#/definitions/jsons.CountryName"
                },
                "tld": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "jsons.CountryName": {
            "type": "object",
            "properties": {
                "common": {
                    "type": "string"
                },
                "official": {
                    "type": "string"
                }
            }
        },
        "philippines.HsPUIsAttributes": {
            "type": "object",
            "properties": {
                "ObjectId": {
                    "type": "integer"
                },
                "PUIs": {
                    "type": "integer"
                },
                "hf": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "philippines.StatsAttributes": {
            "type": "object",
            "properties": {
                "ObjectId": {
                    "type": "integer"
                },
                "PUIs": {
                    "type": "object"
                },
                "PUMs": {
                    "type": "object"
                },
                "confirmed": {
                    "type": "integer"
                },
                "day": {
                    "type": "integer"
                },
                "deaths": {
                    "type": "integer"
                },
                "recovered": {
                    "type": "integer"
                },
                "tests": {
                    "type": "object"
                }
            }
        },
        "rest.BasicError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "who.Data": {
            "type": "object",
            "properties": {
                "case_daily_increase": {
                    "type": "number"
                },
                "date": {
                    "type": "number"
                },
                "death_daily_increase": {
                    "type": "number"
                },
                "region": {
                    "type": "string"
                },
                "total_cases": {
                    "type": "number"
                },
                "total_deaths": {
                    "type": "number"
                }
            }
        },
        "who.HistoryData": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "country_info": {
                    "type": "object",
                    "$ref": "#/definitions/jsons.Country"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/who.Data"
                    }
                }
            }
        },
        "worldometer.CountryStats": {
            "type": "object",
            "properties": {
                "active_cases": {
                    "type": "integer"
                },
                "cases_per_one_million": {
                    "type": "number"
                },
                "country": {
                    "type": "string"
                },
                "country_info": {
                    "type": "object"
                },
                "deaths_per_one_million": {
                    "type": "number"
                },
                "new_cases": {
                    "type": "integer"
                },
                "new_deaths": {
                    "type": "integer"
                },
                "serious_critical": {
                    "type": "integer"
                },
                "tests_per_one_million": {
                    "type": "number"
                },
                "total_cases": {
                    "type": "integer"
                },
                "total_deaths": {
                    "type": "integer"
                },
                "total_recovered": {
                    "type": "integer"
                },
                "total_tests": {
                    "type": "integer"
                }
            }
        },
        "worldometer.StatesStats": {
            "type": "object",
            "properties": {
                "active_cases": {
                    "type": "integer"
                },
                "cases_per_one_million": {
                    "type": "number"
                },
                "deaths_per_one_million": {
                    "type": "number"
                },
                "new_cases": {
                    "type": "integer"
                },
                "new_deaths": {
                    "type": "integer"
                },
                "serious_critical": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "state_info": {
                    "type": "object"
                },
                "tests_per_one_million": {
                    "type": "number"
                },
                "total_cases": {
                    "type": "integer"
                },
                "total_deaths": {
                    "type": "integer"
                },
                "total_recovered": {
                    "type": "integer"
                },
                "total_tests": {
                    "type": "integer"
                }
            }
        }
    }
}